backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
google-generativeai==0.3.2
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0